requirements.txt
# See more documentation in /docs/libraries.md
# This file is here because many Platforms as a Service look for
# requirements.txt in the root directory of a project.
# PBR should always appear first
pbr>=1.8


# Current framework
# https://docs.djangoproject.com/en/1.8/
Django==1.8.14
djangorestframework-csv==2.0.0
django-cors-headers==1.3.1
django-redis==4.7.0
oslo.concurrency==3.16.0
pylibmc==1.5.1
pyjwt==1.4.2
pycrypto==2.6.1
Jinja2==2.9.6
python-dateutil==2.7.3
redis==2.10.6
#celery==4.2.1
#eventlet
#retrying

# python mysql client
# yum install python-devel mysql-devel gcc
MySQL-python==1.2.5


# portal
portal-core>=100.1.0.dev,<100.2.0
portal-rest>=100.1.0.dev,<100.2.0

